Upgrade gradle wrapper to 8.6
Rename project to Click Translate Baidu

Change logo

Upgrade dependencies

Replace dependencies from embed to include
