+++ 1.19.3-1.20.1-fabric/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'fabric-loom' version '1.2-SNAPSHOT'
+    id 'fabric-loom' version '1.5-SNAPSHOT'
     id 'maven-publish'
 }
 
@@ -7,11 +7,6 @@ version = project.mod_version
 group = project.maven_group
 
 repositories {
-    // Add repositories to retrieve artifacts from in here.
-    // You should only use this when depending on other mods because
-    // Loom adds the essential maven repositories to download Minecraft and libraries from automatically.
-    // See https://docs.gradle.org/current/userguide/declaring_repositories.html
-    // for more information about repositories.
     maven {
         url 'https://maven.aliyun.com/repository/public'
     }
@@ -19,15 +14,6 @@ repositories {
     maven {
         url 'https://maven.terraformersmc.com/releases/'
     }
-
-    maven {
-        url 'https://repo.carm.cc/repository/maven-public/'
-    }
-}
-
-configurations {
-    embed
-    api.extendsFrom(embed)
 }
 
 dependencies {
@@ -38,9 +24,12 @@ dependencies {
 
     // Fabric API. This is technically optional, but you probably want it anyway.
     modImplementation "net.fabricmc.fabric-api:fabric-api:${project.fabric_version}"
-    modImplementation 'com.terraformersmc:modmenu:7.1.0'
+    modImplementation 'com.terraformersmc:modmenu:7.2.2'
 
-    embed 'cc.carm.lib:easyconfiguration-hocon:3.7.0'
+    include(api('cc.carm.lib:easyconfiguration-core:3.9.1'))
+	include(api('cc.carm.lib:yamlcommentwriter:1.0.1'))
+	include(api('org.bspfsystems:yamlconfiguration:2.0.0'))
+	include(api('cc.carm.lib:easyconfiguration-yaml:3.9.1'))
 }
 
 processResources {
@@ -81,10 +70,7 @@ java {
 }
 
 jar {
-    from("LICENSE")
-    from configurations.embed.asFileTree.files.collect {
-        it.isDirectory() ? it : zipTree(it)
-    }
+    from 'LICENSE'
 }
 
 // configure the maven publication
