+++ 1.19.3-1.20.1-fabric/gradle.properties
@@ -4,14 +4,14 @@ org.gradle.jvmargs=-Xmx1G
 # Fabric Properties
 	# check these on https://modmuss50.me/fabric.html
 	minecraft_version=1.20.1
-	yarn_mappings=1.20.1+build.9
-	loader_version=0.14.21
+	yarn_mappings=1.20.1+build.10
+	loader_version=0.15.6
 
 # Mod Properties
-	mod_version = 4.1
-	maven_group = online.flowerinsnow
-	archives_base_name = clicktranslate
+	mod_version = 4.2.0
+	maven_group = online.flowerinsnow.clicktranslatebaidu
+	archives_base_name = click-translate-baidu
 
 # Dependencies
 	# check this on https://modmuss50.me/fabric.html
-	fabric_version=0.85.0+1.20.1
+	fabric_version=0.91.0+1.20.1
