--- 1.19.3-1.20.1-fabric/src/main/java/online/flowerinsnow/clicktranslate/client/exception/TranslateException.java
@@ -1,19 +1,0 @@
-package online.flowerinsnow.clicktranslate.client.exception;
-
-public class TranslateException extends RuntimeException {
-    public TranslateException() {
-        super();
-    }
-
-    public TranslateException(String message) {
-        super(message);
-    }
-
-    public TranslateException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public TranslateException(Throwable cause) {
-        super(cause);
-    }
-}
