--- 1.19.3-1.20.1-fabric/src/main/java/online/flowerinsnow/clicktranslate/client/util/MessageUtils.java
@@ -1,14 +1,0 @@
-package online.flowerinsnow.clicktranslate.client.util;
-
-public abstract class MessageUtils {
-    private MessageUtils() {
-    }
-
-    public static String parseColour(String message) {
-        return message.replace("&", "§").replace("§§", "&");
-    }
-
-    public static String removeColour(String message) {
-        return message.replaceAll("§.", "");
-    }
-}
