+++ 1.19.3-1.20.1-fabric/src/main/java/online/flowerinsnow/clicktranslatebaidu/client/exception/TranslateException.java
@@ -1,0 +1,23 @@
+package online.flowerinsnow.clicktranslatebaidu.client.exception;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+
+@Environment(EnvType.CLIENT)
+public class TranslateException extends RuntimeException {
+    public TranslateException() {
+        super();
+    }
+
+    public TranslateException(String message) {
+        super(message);
+    }
+
+    public TranslateException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public TranslateException(Throwable cause) {
+        super(cause);
+    }
+}
