+++ 1.19.3-1.20.1-fabric/src/main/java/online/flowerinsnow/clicktranslatebaidu/client/mixin/MixinChatHud.java
@@ -1,0 +1,29 @@
+package online.flowerinsnow.clicktranslatebaidu.client.mixin;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.client.gui.hud.ChatHud;
+import net.minecraft.client.gui.hud.MessageIndicator;
+import net.minecraft.network.message.MessageSignatureData;
+import net.minecraft.text.Text;
+import net.minecraft.util.ActionResult;
+import online.flowerinsnow.clicktranslatebaidu.client.eci.ChatHudAddMessageCallback;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(ChatHud.class)
+@Environment(EnvType.CLIENT)
+public class MixinChatHud {
+    @Inject(
+            method = "addMessage(Lnet/minecraft/text/Text;Lnet/minecraft/network/message/MessageSignatureData;Lnet/minecraft/client/gui/hud/MessageIndicator;)V",
+            at = @At("HEAD"),
+            cancellable = true
+    )
+    public void addMessage(Text message, MessageSignatureData signature, MessageIndicator indicator, CallbackInfo ci) {
+        if (ChatHudAddMessageCallback.EVENT.invoker().onAddMessage(message, signature, indicator) == ActionResult.FAIL) {
+            ci.cancel();
+        }
+    }
+}
