+++ 1.19.3-1.20.1-fabric/src/main/java/online/flowerinsnow/clicktranslatebaidu/client/object/NoObfuscation.java
@@ -1,0 +1,16 @@
+package online.flowerinsnow.clicktranslatebaidu.client.object;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@NoObfuscation
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Environment(EnvType.CLIENT)
+public @interface NoObfuscation {
+}
