+++ 1.19.3-1.20.1-fabric/src/main/resources/click-translate-baidu.mixins.json
@@ -1,0 +1,13 @@
+{
+  "required": true,
+  "minVersion": "0.8",
+  "package": "online.flowerinsnow.clicktranslatebaidu.client.mixin",
+  "compatibilityLevel": "JAVA_17",
+  "client": [
+    "MixinScreen",
+    "MixinChatHud"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
