--- 1.19.3-1.20.1-fabric/src/main/resources/clicktranslate.mixins.json
@@ -1,15 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "online.flowerinsnow.clicktranslate.client.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [
-  ],
-  "client": [
-    "MixinScreen",
-    "MixinChatHud"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
